backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
python-dotenv==1.0.0
werkzeug==3.0.1
